MIGENG-138 Fixed Agenda Groups focusing order
